requirements.txt
bottle==0.12.19
bottle-flash2==0.0.3
cached-property==1.5.2
charset-normalizer==2.0.9
Deprecated==1.2.13
idna==3.3
importlib-metadata==4.10.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
prompt-toolkit==3.0.26
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
typing-extensions==4.0.1
urllib3==1.26.7
python-gitlab==3.15.0
